app/requirements.txt
Django==4.1.3
Pillow==9.3.0
django-bootstrap5==22.2
djangorestframework==3.14.0
psycopg2-binary==2.9.3
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
gunicorn==21.2.0